Add Inspect failure tests
